Fixed session error

+++ Pages/index.php
@@ -1,8 +1,13 @@
-<?php require "header.php"; ?>
+<?php require "header.php";
+
+var_dump($_SESSION);
+?>
 
             <div id="right">
                 <div class="section">
-                    <div class="left">
+                    <?php
+                    if(isset($_SESSION['id']) && $_SESSION['isAdmin']==1){
+                        echo '<div class="left">
                         <h1>Add user</h1>
                         <form method="post" action="add_user.php">
 
@@ -13,7 +18,11 @@
                             <input type="text" name="isAdmin" placeholder="Admin"/>
                             <input type="submit" value="Save"/>
                         </form>
-                    </div>
+                    </div>';
+                    }
+
+                    ?>
+
                     <div class="right">
                         <?php
 
@@ -30,7 +39,7 @@
                                 <td>Password</td>
                                 <td>Admin</td>";
 
-                                if(isset($_SESSION['id'])){
+                                if(isset($_SESSION['id']) && $_SESSION['isAdmin']==1){
                                     echo "<td>Edit</td><td>Delete</td>";
                                 }
                            echo "</tr>";
@@ -38,7 +47,7 @@
                             foreach($users as $key=>$value){
                                     echo "<tr>";
                                         echo '<td>' . $users[$key]->id . "</td><td>" . $users[$key]->firstName . "</td><td>" . $users[$key]->lastName . "</td><td>" . $users[$key]->username . "</td><td>" . $users[$key]->password . '</td><td>'. $users[$key]->isAdmin . '</td>';
-                                        if(isset($_SESSION['id'])){
+                                        if(isset($_SESSION['id']) && $_SESSION['isAdmin']==1){
                                             echo '<td><a href="edit_user.php?id=' . $users[$key]->id . '" ><img src="../img/file_edit.png" width="30px"></a> </td><td>' . ' <a href="delete_user.php?id=' . $users[$key]->id . '" ><img src="../img/file_delete.png" width="30px"></a></td>';
                                         }
                                     echo "</tr>";
@@ -49,7 +58,9 @@
                     </div>
                 </div>
                 <div class="section">
-                    <div class="left">
+                    <?php
+                    if(isset($_SESSION['id']) && $_SESSION['isAdmin']==1){
+                    echo '<div class="left">
                         <h1>Add Contact</h1>
                         <form method="post" action="add_contact.php">
 
@@ -58,7 +69,8 @@
                             <input type="text" name="userId" placeholder="User ID"/>
                             <input type="submit" value="Save"/>
                         </form>
-                    </div>
+                    </div>';}
+                    ?>
                     <div class="right">
                         <?php
 
@@ -75,14 +87,14 @@
                             <td>First name</td>
                             <td>Last name</td>
                             <td>User ID</td>";
-                            if(isset($_SESSION['id'])){
+                            if(isset($_SESSION['id']) && $_SESSION['isAdmin']==1){
                                 echo "<td>Edit</td><td>Delete</td>";
                             }
                         echo "</tr>";
                         foreach($contacts as $key=>$value){
                             echo "<tr>";
                             echo '<td>' . $contacts[$key]->id . "</td><td>" . $contacts[$key]->firstName . "</td><td>" . $contacts[$key]->lastName . "</td><td>" . $contacts[$key]->userId . "</td>";
-                            if(isset($_SESSION['id'])){
+                            if(isset($_SESSION['id']) && $_SESSION['isAdmin']==1){
                                 echo  '<td><a href="edit_contact.php?id=' . $contacts[$key]->id . '" ><img src="../img/file_edit.png" width="30px"></a> </td><td>' . ' <a href="delete_contact.php?id=' . $contacts[$key]->id . '" ><img src="../img/file_delete.png" width="30px"></a></td>';
                             }
                             echo "</tr>";
@@ -94,7 +106,9 @@
                     </div>
                 </div>
                 <div class="section">
-                    <div class="left">
+                    <?php
+                    if(isset($_SESSION['id']) && $_SESSION['isAdmin']==1) {
+                        echo ' <div class="left">
                         <h1>Add phone</h1>
                         <form method="post" action="add_phone.php">
 
@@ -103,7 +117,9 @@
                             <input type="text" name="contactId" placeholder="Contact ID"/>
                             <input type="submit" value="Save"/>
                         </form>
-                    </div>
+                    </div>';
+                    }
+                    ?>
                     <div class="right">
                         <?php
 
@@ -119,13 +135,17 @@
                             <td>ID</td>
                             <td>Type</td>
                             <td>Number</td>
-                            <td>Contact ID</td>
-                            <td>Edit</td>
-                            <td>Delete</td>
-                        </tr>";
+                            <td>Contact ID</td>";
+                            if(isset($_SESSION['id']) && $_SESSION['isAdmin']==1){
+                                echo "<td>Edit</td><td>Delete</td>";
+                            }
+                        echo "</tr>";
                         foreach($phones as $key=>$value){
                             echo "<tr>";
-                            echo '<td>' . $phones[$key]->id . "</td><td>" . $phones[$key]->type . "</td><td>" . $phones[$key]->number . "</td><td>" . $phones[$key]->contactId . "</td>" . '<td><a href="edit_phone.php?id=' . $phones[$key]->id . '" ><img src="../img/file_edit.png" width="30px"></a> </td><td>' . ' <a href="delete_phone.php?id=' . $phones[$key]->id . '" ><img src="../img/file_delete.png" width="30px"></a></td>';
+                            echo '<td>' . $phones[$key]->id . "</td><td>" . $phones[$key]->type . "</td><td>" . $phones[$key]->number . "</td><td>" . $phones[$key]->contactId . "</td>";
+                            if(isset($_SESSION['id']) && $_SESSION['isAdmin']==1){
+                                echo '<td><a href="edit_phone.php?id=' . $phones[$key]->id . '" ><img src="../img/file_edit.png" width="30px"></a> </td><td>' . ' <a href="delete_phone.php?id=' . $phones[$key]->id . '" ><img src="../img/file_delete.png" width="30px"></a></td>';
+                            }
                             echo "</tr>";
                         }
